Tratamento de Erros

+++ appnovo/src/main/java/br/com/appnovo/repository/PreMedicoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.appnovo.model.PreMedico;
 
 
-public interface PreMedicoRepository extends JpaRepository<PreMedico, Integer> {
+public interface PreMedicoRepository extends JpaRepository<PreMedico, Long> {
 
 	List<PreMedico> findByRegistro(String registro);
